+++ src/core.rs
@@ -1,4 +1,5 @@
 pub mod bus;
+pub mod connection;
 pub mod message;
-pub mod persistence;
+pub mod thread;
 pub mod value;