+++ src/services.rs
@@ -2,6 +2,7 @@ use crate::prelude::*;
 use crate::settings::{Service, Settings};
 use lazy_static::lazy_static;
 use reqwest::header::{HeaderMap, HeaderValue};
+use reqwest::Method;
 use std::time::Duration;
 
 pub mod buienradar;
@@ -9,6 +10,7 @@ pub mod clock;
 pub mod db;
 pub mod openweather;
 pub mod rhai;
+pub mod ring;
 pub mod solar;
 pub mod tado;
 pub mod telegram;
@@ -20,7 +22,7 @@ lazy_static! {
 }
 
 /// Spawn all the configured services.
-pub fn spawn_all(settings: &Settings, service_ids: &Option<Vec<String>>, bus: &mut Bus) -> Result<()> {
+pub fn spawn_all(settings: &Settings, service_ids: &Option<Vec<String>>, bus: &mut Bus, db: &Connection) -> Result {
     for (service_id, service) in settings.services.iter() {
         if let Some(service_ids) = service_ids {
             if !service_ids.contains(service_id) {
@@ -42,6 +44,7 @@ pub fn spawn_all(settings: &Settings, service_ids: &Option<Vec<String>>, bus: &m
                 Service::Tado(tado) => tado.spawn(service_id, bus),
                 Service::Telegram(telegram) => telegram.spawn(service_id, bus),
                 Service::YouLess(youless) => youless.spawn(service_id, bus),
+                Service::Ring(ring) => ring.spawn(service_id, bus, db),
             }
         } {
             error!("Failed to spawn `{}`: {}", service_id, error.to_string());
@@ -50,6 +53,7 @@ pub fn spawn_all(settings: &Settings, service_ids: &Option<Vec<String>>, bus: &m
     Ok(())
 }
 
+/// Builds an HTTP client to use with a service.
 fn build_client() -> Result<Client> {
     let mut headers = HeaderMap::new();
     headers.insert(reqwest::header::USER_AGENT, HeaderValue::from_static(USER_AGENT));
@@ -61,6 +65,21 @@ fn build_client() -> Result<Client> {
         .build()?)
 }
 
+/// Deserializes a Unix time into `DateTime<Local>`.
 fn deserialize_timestamp<'de, D: Deserializer<'de>>(deserializer: D) -> Result<DateTime<Local>, D::Error> {
     Ok(Local.timestamp(i64::deserialize(deserializer)?, 0))
+}
+
+/// Generic function to call a JSON API.
+fn call_json_api<U, R>(method: Method, access_token: &str, url: U) -> Result<R>
+where
+    U: AsRef<str> + std::fmt::Display,
+    R: DeserializeOwned,
+{
+    debug!("Calling {}…", url);
+    Ok(CLIENT
+        .request(method, url.as_ref())
+        .header("Authorization", format!("Bearer {}", access_token))
+        .send()?
+        .json()?)
 }