+++ src/services/db.rs
@@ -1,7 +1,6 @@
 //! System database service.
 
 use crate::prelude::*;
-use std::thread;
 use std::time::Duration;
 
 const INTERVAL: Duration = Duration::from_secs(60);
@@ -9,20 +8,12 @@ const INTERVAL: Duration = Duration::from_secs(60);
 pub struct Db;
 
 impl Db {
-    pub fn spawn(self, service_id: String, bus: &mut Bus, db: Connection) -> Result<()> {
+    pub fn spawn(self, service_id: String, bus: &mut Bus, db: Connection) -> Result {
         let tx = bus.add_tx();
-
-        thread::Builder::new().name(service_id).spawn(move || loop {
-            if let Err(error) = self.loop_(&db, &tx) {
-                error!("Failed to refresh the sensors: {}", error.to_string());
-            }
-            thread::sleep(INTERVAL);
-        })?;
-
-        Ok(())
+        spawn_service_loop(service_id, INTERVAL, move || self.loop_(&db, &tx))
     }
 
-    fn loop_(&self, db: &Connection, tx: &Sender) -> Result<()> {
+    fn loop_(&self, db: &Connection, tx: &Sender) -> Result {
         tx.send(
             Message::new("db::size")
                 .value(Value::DataSize(db.select_size()?))