+++ src/services/solar.rs
@@ -1,7 +1,5 @@
-use crate::core::message::Type;
 use crate::prelude::*;
 use spa::{calc_sunrise_and_set, SunriseAndSet};
-use std::thread;
 use std::time::Duration;
 
 /// Emits durations to and after sunrise and sunset.
@@ -33,60 +31,60 @@ const fn default_interval_millis() -> u64 {
 }
 
 impl Solar {
-    pub fn spawn(self, service_id: String, bus: &mut Bus) -> Result<()> {
+    pub fn spawn(self, service_id: String, bus: &mut Bus) -> Result {
         let tx = bus.add_tx();
-        let interval = Duration::from_millis(self.interval_millis);
 
-        thread::Builder::new().name(service_id.clone()).spawn(move || loop {
-            let now = Utc::now();
-            match calc_sunrise_and_set(now, self.secrets.latitude, self.secrets.longitude) {
-                Ok(SunriseAndSet::Daylight(sunrise, sunset)) => {
-                    if now < sunrise {
-                        Message::new(format!("{}::before::sunrise", service_id))
-                            .sensor_title("Time Before Sunrise")
+        spawn_service_loop(
+            service_id.clone(),
+            Duration::from_millis(self.interval_millis),
+            move || {
+                let now = Utc::now();
+                match calc_sunrise_and_set(now, self.secrets.latitude, self.secrets.longitude)? {
+                    SunriseAndSet::Daylight(sunrise, sunset) => {
+                        if now < sunrise {
+                            Message::new(format!("{}::before::sunrise", service_id))
+                                .sensor_title("Time Before Sunrise")
+                                .optional_location(self.location.clone())
+                                .value(Value::Duration((sunrise - now).num_seconds() as f64))
+                                .send_and_forget(&tx);
+                        }
+                        if now < sunset {
+                            Message::new(format!("{}::before::sunset", service_id))
+                                .sensor_title("Time Before Sunset")
+                                .optional_location(self.location.clone())
+                                .value(Value::Duration((sunset - now).num_seconds() as f64))
+                                .send_and_forget(&tx);
+                        }
+                        if sunrise < now {
+                            Message::new(format!("{}::after::sunrise", service_id))
+                                .sensor_title("Time After Sunrise")
+                                .optional_location(self.location.clone())
+                                .value(Value::Duration((now - sunrise).num_seconds() as f64))
+                                .send_and_forget(&tx);
+                        }
+                        if sunset < now {
+                            Message::new(format!("{}::after::sunset", service_id))
+                                .sensor_title("Time After Sunset")
+                                .optional_location(self.location.clone())
+                                .value(Value::Duration((now - sunset).num_seconds() as f64))
+                                .send_and_forget(&tx);
+                        }
+                    }
+                    SunriseAndSet::PolarDay => {
+                        Message::new(format!("{}::polar::day", service_id))
+                            .type_(MessageType::ReadNonLogged)
                             .optional_location(self.location.clone())
-                            .value(Value::Duration((sunrise - now).num_seconds() as f64))
                             .send_and_forget(&tx);
                     }
-                    if now < sunset {
-                        Message::new(format!("{}::before::sunset", service_id))
-                            .sensor_title("Time Before Sunset")
+                    SunriseAndSet::PolarNight => {
+                        Message::new(format!("{}::polar::night", service_id))
+                            .type_(MessageType::ReadNonLogged)
                             .optional_location(self.location.clone())
-                            .value(Value::Duration((sunset - now).num_seconds() as f64))
-                            .send_and_forget(&tx);
-                    }
-                    if sunrise < now {
-                        Message::new(format!("{}::after::sunrise", service_id))
-                            .sensor_title("Time After Sunrise")
-                            .optional_location(self.location.clone())
-                            .value(Value::Duration((now - sunrise).num_seconds() as f64))
-                            .send_and_forget(&tx);
-                    }
-                    if sunset < now {
-                        Message::new(format!("{}::after::sunset", service_id))
-                            .sensor_title("Time After Sunset")
-                            .optional_location(self.location.clone())
-                            .value(Value::Duration((now - sunset).num_seconds() as f64))
                             .send_and_forget(&tx);
                     }
                 }
-                Ok(SunriseAndSet::PolarDay) => {
-                    Message::new(format!("{}::polar::day", service_id))
-                        .type_(Type::ReadNonLogged)
-                        .optional_location(self.location.clone())
-                        .send_and_forget(&tx);
-                }
-                Ok(SunriseAndSet::PolarNight) => {
-                    Message::new(format!("{}::polar::night", service_id))
-                        .type_(Type::ReadNonLogged)
-                        .optional_location(self.location.clone())
-                        .send_and_forget(&tx);
-                }
-                Err(error) => error!("Failed to calculate sunrise and sunset: {}", error.to_string()),
-            }
-            thread::sleep(interval);
-        })?;
-
-        Ok(())
+                Ok(())
+            },
+        )
     }
 }